Processing spectral data for tests of the spectral variability hypothesis

Drone data is complex and extracting spectral diversity values requires several processing steps. This package is designed to be part of the workflow for analyzing these images

Functions

Drone imagery processing

create_masked_raster

create_multiband_image

extract_pixel_values

Spectral diversity

calculate_cv

calculate_sv

calculate_chv_nopca

calculate_spectral_metrics

On the ground diversity

calculate_field_diversity

References

If you have any questions, comments or suggestions, please submit an issue at our GitHub repository

See also

Author

Maintainer: Adele Gemmell adelegemmell@hotmail.com (ORCID)

Authors:

Adele Gemmell adelegemmell@hotmail.com (ORCID)

Will Cornwell w.cornwell@unsw.edu.au (ORCID) [contributor]

Other contributors:

Adrian Fisher adrian.fisher@unsw.edu.au (ORCID) [contributor]